+++ src/base/image_base.c
@@ -125,7 +125,7 @@ ret_t image_base_set_prop(widget_t* widget, const char* name, const value_t* v) 
   return RET_NOT_FOUND;
 }
 
-ret_t image_base_destroy(widget_t* widget) {
+ret_t image_base_on_destroy(widget_t* widget) {
   image_base_t* image = IMAGE_BASE(widget);
   return_value_if_fail(image != NULL, RET_BAD_PARAMS);