+++ src/base/widget.h
@@ -64,8 +64,8 @@ typedef ret_t (*widget_get_prop_default_value_t)(widget_t* widget, const char* n
 typedef ret_t (*widget_set_prop_t)(widget_t* widget, const char* name, const value_t* v);
 typedef widget_t* (*widget_find_target_t)(widget_t* widget, xy_t x, xy_t y);
 typedef widget_t* (*widget_create_t)(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h);
+typedef ret_t (*widget_on_destroy_t)(widget_t* widget);
 typedef ret_t (*widget_recycle_t)(widget_t* widget);
-typedef ret_t (*widget_destroy_t)(widget_t* widget);
 
 typedef struct _widget_vtable_t {
   uint32_t size;
@@ -109,7 +109,7 @@ typedef struct _widget_vtable_t {
   widget_on_event_t on_event;
   widget_find_target_t find_target;
   widget_recycle_t recycle;
-  widget_destroy_t destroy;
+  widget_on_destroy_t on_destroy;
 } widget_vtable_t;
 
 /**