+++ src/base/window_manager.c
@@ -591,7 +591,7 @@ static ret_t window_manager_set_prop(widget_t* widget, const char* name, const v
   return RET_NOT_FOUND;
 }
 
-static ret_t window_manager_destroy(widget_t* widget) {
+static ret_t window_manager_on_destroy(widget_t* widget) {
   window_manager_t* wm = WINDOW_MANAGER(widget);
 
   TKMEM_FREE(wm->cursor);
@@ -607,7 +607,7 @@ static const widget_vtable_t s_window_manager_vtable = {
     .on_paint_children = window_manager_on_paint_children,
     .on_remove_child = wm_on_remove_child,
     .find_target = window_manager_find_target,
-    .destroy = window_manager_destroy};
+    .on_destroy = window_manager_on_destroy};
 
 static ret_t wm_on_locale_changed(void* ctx, event_t* e) {
   widget_t* widget = WIDGET(ctx);