+++ src/widgets/popup.c
@@ -117,7 +117,7 @@ static const widget_vtable_t s_popup_vtable = {.size = sizeof(popup_t),
                                                .on_paint_self = window_base_on_paint_self,
                                                .on_paint_begin = window_base_on_paint_begin,
                                                .on_paint_end = window_base_on_paint_end,
-                                               .destroy = window_base_destroy};
+                                               .on_destroy = window_base_on_destroy};
 
 widget_t* popup_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h) {
   popup_t* popup = TKMEM_ZALLOC(popup_t);